==> Evidence/Hannah Harrison/02-Data.Rmd <==
---
title: "02 - Data"
author: "Hannah Harrison"
date: "2022-11-11"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

### Library requirements 

```{r}
if (!require("fs")) install.packages("fs")
library("fs")
if (!require("magrittr")) install.packages("magrittr")
library("magrittr")
```

To obtain this dataset in a convenient format, we will download it and process it into a standard form (containing headers). We will then further remove 10% of the 'kddcup.data_10_percent.gz' dataset at random, so that the same test dataset can be used for all models.

### Get the data

First create a directory for it to go in:
```{r}
rawdatadir=path_wd("..","data")
if(!dir.exists(rawdatadir)) dir.create(rawdatadir,recursive = TRUE)
```

Download it
```{r}
download.file('http://kdd.ics.uci.edu/databases/kddcup99/kddcup.data_10_percent.gz', path_wd("..","data","kddcup.data_10_percent.gz"))

download.file('http://kdd.ics.uci.edu/databases/kddcup99/kddcup.testdata.unlabeled_10_percent.gz', path_wd("..","data","kddcup.testdata.unlabeled_10_percent.gz"))

download.file('http://kdd.ics.uci.edu/databases/kddcup99/corrected.gz', path_wd("..","data","corrected.gz"))

download.file('http://kdd.ics.uci.edu/databases/kddcup99/kddcup.names', path_wd("..","data","kddcup.names"))

```
### Add column names to datasets

Read in data:
```{r}
cnames = read.table(path_wd("..","data","kddcup.names"),
                    as.is=T, skip=1, sep=":")

kdd_10 = read.csv(path_wd("..","data","kddcup.data_10_percent.gz"))
labeled_test = read.csv(path_wd("..","data","corrected.gz"))
unlabeled_test = read.csv(path_wd("..","data","kddcup.testdata.unlabeled_10_percent.gz"))


```
Add column names. Note that the 'unlabeled' dataset has no 'normal' column.

```{r}

datafiles = list(kdd_10, labeled_test, unlabeled_test)

colnames(kdd_10)=c(cnames[,1],"normal")
colnames(labeled_test)=c(cnames[,1],"normal")
colnames(unlabeled_test) = c(cnames[,1])


```


Check that this has worked:
```{r}
head(kdd_10)
head(labeled_test)
head(unlabeled_test)
```

### Preprocessing

All columns apart from normal need to be converted to factor or a numeric.
```{r}
cols.fac <- c(2,3,4,7,12,21,22)
for (file in datafiles) {
  sapply(file[,cols.fac],as.factor)
}

v <- 1:41
cols.num <- v[!(v %in% cols.fac)]

for (file in datafiles) {
  sapply(file[,cols.num],as.character)
}
for (file in datafiles) {
  sapply(file[,cols.num],as.numeric)
}

```

We are also predominantly interested in predicting the general attack class, rather than the specific attack type, as the unlabeled test dataset contains attack types which aren't represented in the training set. Hence I will add a column to the datasets specifying the attack class. Within the dataset, there are 4 attack classes; Denial of Service (DoS), Probe, User to Root(U2R), and Remote to Local (R2L).

Here I extract the labels and count how many types we have.


```{r}
attack_types1 <- unique(labeled_test$normal) 
attack_types2 <- unique(kdd_10$normal)
attack_types <- unique(c(attack_types1, attack_types2))
attack_types
length(attack_types1)
length(attack_types2)
length(attack_types)

```
Now I add the column with attack class. 

Defining the classes:

```{r}
DDoS = c('neptune.', 'smurf.', 'pod.', 'teardrop.', 'land.', "back.", "apache2.", "udpstorm.", "processtable.", "mailbomb.")
UR2 = c("buffer_overflow.","loadmodule.", "perl.", "rootkit.", "xterm.", "ps.", "httptunnel.", "sqlattack.", "worm.", "snmpguess.", "spy." )
R2L = c("guess_passwd.", "ftp_write.", "imap.", "phf.", "multihop.", "warezmaster.", "warezclient.", "snmpgetattack.", "named.", "xlock." , "xsnoop.", "sendmail." )
Probe = c("ipsweep.", "portsweep.", "nmap.", "satan." , "saint.", "mscan.")

length(DDoS) + length(UR2) + length(R2L) + length(Probe) + 1
# Correct number of types

```

Creating the class column:

```{r}
kdd_10 <- within(kdd_10, {
  class = 0 
  class[normal %in% c("normal.")] = 'normal'
  class[normal %in% DDoS] = 'DDoS'
  class[normal %in% UR2] = 'UR2'
  class[normal %in% R2L] = 'R2L'
  class[normal %in% Probe] = 'Probe'
})

labeled_test <- within(labeled_test, {
  class = 0 
  class[normal %in% c("normal.")] = 'normal'
  class[normal %in% DDoS] = 'DDoS'
  class[normal %in% UR2] = 'UR2'
  class[normal %in% R2L] = 'R2L'
  class[normal %in% Probe] = 'Probe'
})
```

Check this has worked:
```{r}
unique(kdd_10$class)
unique(labeled_test$class)
```





### Separating the test and training sets from kdd_10

```{r}
n = length(kdd_10[,1])
test_sample = sample(1:n, n/10)

kdd_10_train = kdd_10[!(1:n %in% test_sample), ]
kdd_10_test = kdd_10[test_sample,]
```

Checking we have the correct dimensions:
```{r}
dim(kdd_10_test) # 49402 observations
dim(kdd_10_train) # 444618 observations
```

### Create the processed data output.

```{r}

write.table(kdd_10_train,
            path_wd("..","data","kdd_10_train.tab"),
            quote=FALSE,row.names=FALSE)
write.table(kdd_10_test,
            path_wd("..","data","kdd_10_test.tab"),
            quote=FALSE,row.names=FALSE)
write.table(labeled_test,
            path_wd("..","data","labeled_test.tab"),
            quote=FALSE,row.names=FALSE)
write.table(unlabeled_test,
            path_wd("..","data","unlabeled_test.tab"),
            quote=FALSE,row.names=FALSE)

```
### Summary:

We have now created our data for training and testing, in a simple headed tabular format (which is space separated). This can be easily read into either R or Python.


### Reference:
List of attack types which fall into each class taken from: https://www.researchgate.net/figure/The-distribution-of-attack-types-in-the-KDD-Cup-99-and-NSL-KDD_tbl1_298318568
